refactor!: move static qualified name methods to `local\metric_record`
